refactor: extract Jimaku matching helpers
